My first (big) commit : linked frequency bugfix

+++ LocalRest/DecisionMaker/AdaptiveDecisions/Src/FreqSelector.cpp
@@ -26,11 +26,11 @@
 FreqSelector::FreqSelector (unsigned int nbFreqs)
 {
    this->nbFreqs = nbFreqs;
-   this->coeffs = new unsigned int [nbFreqs];
+   this->coeffs = new float [nbFreqs];
 
    for (unsigned int i = 0; i < nbFreqs; i++)
    {
-      this->coeffs [i] = 0;
+      this->coeffs [i] = 0.;
    }
 }
